Add likes feature (#2)
* Completed step 2 on part one, implemented logout route

* Completed steps 3 & 4 in part one, step 5 in progress

* Steps 5 & 6 complete, implemented a way to edit user profile and fixed homepage to show followed user messages

* Completed part 2, added likes feature

* Finished tests for user and message views

+++ app.py
@@ -5,7 +5,9 @@ from flask_debugtoolbar import DebugToolbarExtension
 from sqlalchemy.exc import IntegrityError
 
 from forms import UserAddForm, LoginForm, MessageForm, EditUserForm
-from models import db, connect_db, User, Message
+
+from models import db, connect_db, User, Message, Likes
+
 import pdb
 
 CURR_USER_KEY = "curr_user"
@@ -114,6 +116,10 @@ def login():
 def logout():
     """Handle logout of user."""
     # IMPLEMENT THIS
+    do_logout()
+    flash("Logout successful, goodbye!", 'success')
+
+    return redirect(url_for("login"))
 
     do_logout()
     flash("Logout successful, goodbye!", 'success')
@@ -214,7 +220,9 @@ def stop_following(follow_id):
 
 @app.route('/users/profile', methods=["GET", "POST"])
 def profile():
-    """Update profile for current user."""
+
+    """GET- user profile edit form, POST - Update profile for current user."""
+
     # IMPLEMENT THIS
 
     if not g.user:
@@ -224,15 +232,21 @@ def profile():
     form = EditUserForm()
 
     if request.method == 'POST':
+
         if form.validate_on_submit():
-            user = User.authenticate(g.user.username,
-                                     form.password.data)
-            if user:
+
+            if User.check_password(g.user, form.password.data):
+
+                user = g.user
+
                 user.username = form.username.data
                 user.email = form.email.data
                 user.image_url = form.image_url.data
                 user.bio = form.bio.data
                 user.location = form.location.data
+
+                user.header_image_url = form.header_image_url.data
+
 
                 db.session.add(user)
                 db.session.commit()
@@ -241,6 +255,7 @@ def profile():
             else:
                 flash("Incorrect password, profile was not updated.", "danger")
                 return redirect(url_for("homepage"))
+
 
     return render_template("/users/edit.html", user=g.user, form=form)
 
@@ -251,29 +266,43 @@ def delete_user():
 
     if not g.user:
         flash("Access unauthorized.", "danger")
-        return redirect("/")
+        return redirect(url_for("homepage"))
 
     do_logout()
 
     db.session.delete(g.user)
     db.session.commit()
 
-    return redirect("/signup")
+    return redirect(url_for("signup"))
 
 
-##############################################################################
+@app.route('/users/<int:user_id>/likes')
+def user_likes(user_id):
+    """ Show list of likes for this user. """
+
+    if not g.user:
+        flash("Access unauthorized.", "danger")
+        return redirect(url_for("homepage"))
+
+    user = User.query.get_or_404(user_id)
+    sorted_likes = user.sort_liked_messages()
+
+    return render_template('/users/likes.html', user=user, sorted_likes=sorted_likes)
+
+###################################################################
 # Messages routes:
+
 
 @app.route('/messages/new', methods=["GET", "POST"])
 def messages_add():
     """Add a message:
 
-    Show form if GET. If valid, update message and redirect to user page.
+    Show form if GET. If valid, create message and redirect to user page.
     """
 
     if not g.user:
         flash("Access unauthorized.", "danger")
-        return redirect("/")
+        return redirect(url_for("homepage"))
 
     form = MessageForm()
 
@@ -301,13 +330,43 @@ def messages_destroy(message_id):
 
     if not g.user:
         flash("Access unauthorized.", "danger")
-        return redirect("/")
+        return redirect(url_for("homepage"))
 
     msg = Message.query.get(message_id)
+
+    if msg.user_id != g.user.id:
+        flash("Access unauthorized.", "danger")
+        return redirect(url_for("homepage"))
+
     db.session.delete(msg)
     db.session.commit()
 
     return redirect(f"/users/{g.user.id}")
+
+
+@app.route("/users/add_like/<int:msg_id>", methods=["POST"])
+def add_message_to_likes(msg_id):
+    """Add message id and user id to Likes table"""
+    if not g.user:
+        flash("Access unauthorized.", "danger")
+        return redirect(url_for("homepage"))
+
+    liked_msg = Likes.query.filter_by(
+        user_id=g.user.id, message_id=msg_id).first()
+
+    if liked_msg:
+        # unlike if message was already liked
+        db.session.delete(liked_msg)
+        db.session.commit()
+        flash("Message has been removed from likes", "success")
+
+    else:
+        liked_msg = Likes(user_id=g.user.id, message_id=msg_id)
+        db.session.add(liked_msg)
+        db.session.commit()
+        flash("Message added to likes", "success")
+
+    return redirect(url_for("homepage"))
 
 
 ##############################################################################
@@ -326,10 +385,13 @@ def homepage():
     if g.user:
         messages = g.user.get_followed_user_messages()
 
+        likes = [like.id for like in g.user.likes]
+
+
         if not messages:
             flash("Messages from people you are following will show up here, use the search bar to find some people to follow!", 'info')
 
-        return render_template('home.html', messages=messages)
+        return render_template('home.html', messages=messages, likes=likes)
 
     else:
         return render_template('home-anon.html')
